requirements.txt
asyncio
boto3
numpy
py-cui
pyaudio
amazon-transcribe
python-dotenv